Week-5/requirements.txt
requests>=2.31.0
langchain>=0.1.0
langchain-aws>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.6
boto3>=1.34.72